% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridEZ_fn_v2_1.R
\name{gridEZ}
\alias{gridEZ}
\title{gridEZ: generate a gridded sampling frame}
\usage{
gridEZ(
  population_raster_path,
  strata1_raster_path,
  strata2_raster_path,
  predefined_EZ_size = TRUE,
  EZ_target_size = "medium",
  target_pop_per_EZ = 500,
  max_cells_per_EZ = 900,
  output_path,
  run_ID = "run1",
  ncores = 1
)
}
\arguments{
\item{population_raster_path}{Character string. Full file path to a population raster file. Each pixel/cell should contain a population count. NAs allowed. Cells containing NAs are excluded from EZs.}

\item{strata1_raster_path}{Character string. Full file path to a strata raster file. Each pixel/cell should contain an ID corresponding to a stratum, e.g. administrative unit. NAs allowed. Cells containing NAs are excluded from EZs.}

\item{strata2_raster_path}{Character string. Full file path to a strata raster file. Each pixel/cell should contain an ID corresponding to a stratum. Strata 2 will usually be different to strata 1 depending on the stratification requirements of a survey, e.g. settlement type classification, climate zone, categories of distance to a geographical feature. If only one form of stratification is required, simply repeat the file path for Strata 1 here. NAs allowed. Cells containing NAs are excluded from EZs.}

\item{predefined_EZ_size}{Logical. Default is TRUE. If TRUE, gridEZ will use pre-set parameters for target_pop_per_EZ & max_cells_per_EZ to construct EZs. Set to FALSE if you want to specify your own target_pop_per_EZ & max_cells_per_EZ.}

\item{EZ_target_size}{Character sting. Considered only if predefined_EZ_size = TRUE. Default is "medium". Can be either "small", "medium" (or "med") or "large". For "small" target_pop_per_EZ = 75 and max_cells_per_EZ = 100; for "medium" target_pop_per_EZ = 500 and max_cells_per_EZ = 900; for "large" target_pop_per_EZ = 1200 and max_cells_per_EZ = 2500}

\item{target_pop_per_EZ}{Numeric, single number. Considered only if predefined_EZ_size = FALSE. Default is 500. This should be specified if you would like to define your own target population per EZ.}

\item{max_cells_per_EZ}{Numeric, single number. Considered only if predefined_EZ_size = FALSE. Default is 900 raster cells. This should be specified if you would like to define your own maximum cells per EZ. Note, this is not a hard maximum to allow flexibility in EZs reaching the target_pop_per_EZ.}

\item{output_path}{Character string. File path for the location you would like output files to be saved.}

\item{run_ID}{Character string. Default is "_run1". Text to be added to the end of output file names.}

\item{ncores}{Numeric, single number. Default is 1. The number of computer cores you would like for processing. Recommend using your total number of cores minus 1. Use parallel::detectCores() to find the total number of cores available on your computer.}
}
\value{
Raster and csv saved to file. Raster contain the EZ ID for each grid cell. The csv contains columns of EZ IDs ("EZ_IDs"), population per EZ ("total_pop") and number of grid cells per EZ ("N").
}
\description{
gridEZ generates a sampling frame consisting of gridded enumeration zones (EZs). EZs are constructed based on a target population count per EZ whilst being restricted by a maximum geographic size.
gridEZ uses gridded population counts and two stratification datasets that must all be rasters in the same coordinate reference system (CRS) with the same spatial extents.
For household or population surveys, the stratification datasets could be settlement types and administrative units, for example. If one or no stratification is needed, the "strata1" and/or "strata2" rasters (specified in the function parameters) can simply contain one value everywhere. Note that large areas for a single stratum is not advised as it will require a high level of computation and likely lead to very slow running time and/or a memory error.
}
